requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-allauth==65.13.0
django-humanize==0.1.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
humanize==4.14.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.11.0
django-storages==1.13.1
# Relax boto3 pin so pip/railway can pick a compatible build for their environment.
# Use a broad but safe range: any 1.x boto3 (>=1.26, <2.0).
boto3>=1.26,<2.0
cloudinary==1.33.0
# Relax cloudinary storage pin so Railway (Python 3.11) can pick a compatible release.
# Some newer releases may require Python >=3.12; allow any 0.x/1.x release compatible with 3.11
django-cloudinary-storage>=0.3.0,<2.0

# Background task queue
celery==5.3.1
# Use a broad, compatible redis client range that is available on PyPI for Python 3.11
redis>=4.6,<5.0
